chore: adding viewcontroller extension to keep screens switch tidy
