+++ GDSCommon-Demo/GDSCommon-Demo/Extensions/ViewControllerInitialiser+Extensions.swift
@@ -1,0 +1,16 @@
+import GDSCommon
+import UIKit
+
+extension DatePickerScreenViewController {
+    convenience init() {
+        let datePickerVM = MockDatePickerViewModel(selectedDate: Date(),
+                                    minDate: nil,
+                                    maxDate: nil)
+        
+        let viewModel = MockDatePickerScreenViewModel(title: "",
+                                          datePickerViewModel: datePickerVM,
+                                          buttonViewModel: MockButtonViewModel(title: "Action Button", shouldLoadOnTap: false, action: {}))
+        
+        self.init(viewModel: viewModel)
+    }
+}
